doctorbot/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
mistralai==1.0.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
